move AlphaFoldParams out to parameters.py
