+++ alphafold/partition.py
@@ -4,27 +4,7 @@ from alphafold.secstruct import *
 from alphafold.explicit_recursions import *
 from alphafold.dynamic_programming import *
 from alphafold.backtrack import mfe
-
-##################################################################################################
-class AlphaFoldParams:
-    '''
-    Parameters that Define the statistical mechanical model for RNA folding
-    '''
-    def __init__( self ):
-        # Seven parameter model
-        self.C_init = 1.0     # Effective molarity for starting each loop (units of M)
-        self.l      = 0.5     # Effective molarity penalty for each linkages in loop (dimensionless)
-        self.Kd_BP  = 0.0002  # Kd for forming base pair (units of M )
-        self.l_BP   = 0.2     # Effective molarity penalty for each base pair in loop (dimensionless)
-        self.C_eff_stacked_pair = 1e4 # Effective molarity for forming stacked pair (units of M)
-        self.K_coax = 100     # coax bonus for contiguous helices (dimensionless). Set to 0 to turn off coax (dimensionless)
-        self.l_coax = 200     # Effective molarity bonus for each coaxial stack in loop. Initial guess: C_eff_stacked_pair / (C_init*l*K_coax)
-        self.C_std = 1.0      # 1 M. drops out in end (up to overall scale factor).
-        self.min_loop_length = 1 # Disallow apical loops smaller than this size (integer)
-        self.allow_strained_3WJ = False # Prevent strained three-way-junctions with two helices coaxially stacked and no spacer nucleotides to other helix.
-
-    def get_variables( self ):
-        return ( self.C_init, self.l, self.Kd_BP, self.l_BP, self.C_eff_stacked_pair, self.K_coax, self.l_coax, self.C_std, self.min_loop_length, self.allow_strained_3WJ )
+from alphafold.parameters import AlphaFoldParams
 
 ##################################################################################################
 def partition( sequences, params = AlphaFoldParams(), circle = False, verbose = False, calc_deriv = False, use_explicit_recursions = True, backtrack = False ):
